split observability components in different files for readability
